filterDefaultValues
