+++ frontoffice/src/resources/Activity/Event/EventList.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { ListBase, ShowButton } from 'react-admin';
 import { Box, useMediaQuery } from '@material-ui/core';
 import { useLocation } from 'react-router';
@@ -10,12 +10,50 @@ import Calendar from '../../../svg/CalendarIcon';
 import MapIcon from '../../../svg/MapIcon';
 import ListIcon from '@material-ui/icons/List';
 import Filter from '../../../commons/Filter';
-import UpcomingFilter from '../../../commons/UpcomingFilter';
+import SparqlFilter from '../../../commons/SparqlFilter';
 import CardsList from '../../../commons/lists/CardsList';
 import EventCard from './EventCard';
 
 const EventList = (props) => {
   const xs = useMediaQuery((theme) => theme.breakpoints.down('xs'), { noSsr: true });
+  
+  const sparqlWhere = useMemo(() => {
+    const now = new Date();
+    return [
+      {
+        type: "bgp",
+        triples: [
+          {
+            "subject": { termType: "Variable", value:"s1" },
+            "predicate": { termType:"NameNode", value: "http://virtual-assembly.org/ontologies/pair#endDate" },
+            "object": { termType:"Variable", value: "endDate" }
+          }
+        ]
+      },{
+      type: "filter",
+      expression:{
+        type: "operation",
+        operator: ">",
+        args:[
+          {
+            termType: "Variable",
+            value: "endDate"
+          },
+          {
+            termType: "Literal",
+            datatype: {
+              termType:"NamedNode",
+              value:"http://www.w3.org/2001/XMLSchema#dateTime"
+            },
+            language: "",
+            // value: "2022-11-17T10:20:13+05:30"
+            value: now.toISOString()
+          }
+        ]
+      }
+    }
+    ]
+  }, []);
 
   // Filter out finished events only for non-calendar view
   const query = new URLSearchParams(useLocation().search);
@@ -23,10 +61,16 @@ const EventList = (props) => {
   const filter = view === 'calendar' ? undefined : { 'pair:hasStatus': process.env.REACT_APP_MIDDLEWARE_URL + 'status/open' };
 
   return (
-    <ListBase perPage={1000} filter={filter} sort={{ field: 'pair:startDate', order: 'ASC' }} {...props}>
+    <ListBase 
+      perPage={1000}
+      filter={filter}
+      filterDefaultValues={{'sparqlWhere': sparqlWhere}}
+      sort={{ field: 'pair:startDate', order: 'ASC' }}
+      {...props}
+    >
       <MultiViewsFilterList
         filters={[
-          <UpcomingFilter initialChecked source="http://virtual-assembly.org/ontologies/pair#endDate" label="N'afficher que les événements à venir" />,
+          <SparqlFilter initialChecked sparqlWhere={sparqlWhere} label="N'afficher que les événements à venir" />,
           <Filter reference="Region" source="pair:hasLocation" inverseSource="pair:locationOf" label="Région" />,
           <Filter reference="Theme" source="pair:hasTopic" inverseSource="pair:topicOf" label="Thématique" />,
           <Filter
